Move implemenation to module
Separate the python code from other files to better handle different mapping projects
