+++ structure_comparer/__init__.py
@@ -1,0 +1,3 @@
+from .compare import compare_profiles
+from .results_dict import gen_mapping_dict
+from .results_html import create_results_html
