+++ structure_comparer/helpers.py
